add bigdecimals

+++ src/main/java/com/conteo/api/models/dtos/ProblemDtoRs.java
@@ -3,10 +3,12 @@ package com.conteo.api.models.dtos;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class ProblemDtoRs {
 
-    private Double resultado;
+    private BigDecimal resultado;
     private String description;
 }
